Remove the deployment tab
